moved cellprofiler.utilities.rules into the cellprofiler package (resolves #1929)
