+++ cellprofiler/modules/filterobjects.py
@@ -37,7 +37,6 @@ import os
 import scipy as sp
 import scipy.ndimage as scind
 from scipy.sparse import coo_matrix
-import traceback
 
 from cellprofiler.modules.identify import FF_PARENT, FF_CHILDREN_COUNT, \
      M_LOCATION_CENTER_X, M_LOCATION_CENTER_Y
@@ -45,17 +44,13 @@ import cellprofiler.image as cpi
 import cellprofiler.module as cpm
 import cellprofiler.object as cpo
 import cellprofiler.setting as cps
-from cellprofiler.setting import YES, NO
 import cellprofiler.measurement as cpmeas
 import cellprofiler.preferences as cpprefs
-import cellprofiler.utilities.rules as cprules
+import cellprofiler.rules as cprules
 from centrosome.outline import outline
-from centrosome.cpmorphology import fixup_scipy_ndimage_result as fix
 from cellprofiler.modules.identify import add_object_count_measurements
 from cellprofiler.modules.identify import add_object_location_measurements
 from cellprofiler.modules.identify import get_object_measurement_columns
-from cellprofiler.preferences import IO_FOLDER_CHOICE_HELP_TEXT
-from cellprofiler.gui.help import RETAINING_OUTLINES_HELP, NAMING_OUTLINES_HELP
 
 '''Minimal filter - pick a single object per image by minimum measured value'''
 FI_MINIMAL = "Minimal"
