+++ cellprofiler/modules/flagimage.py
@@ -26,18 +26,15 @@ modules upon which the flags are based.
 
 import logging
 import os
-import sys
 
 import numpy as np
 
-import cellprofiler.module as cpm
 import cellprofiler.measurement as cpmeas
+import cellprofiler.module as cpm
+import cellprofiler.rules as cprules
 import cellprofiler.setting as cps
-import cellprofiler.utilities.rules as cprules
 import cellprofiler.workspace as cpw
-from cellprofiler.gui.help import USING_METADATA_TAGS_REF, USING_METADATA_HELP_REF
-from cellprofiler.preferences import IO_FOLDER_CHOICE_HELP_TEXT
-from cellprofiler.setting import YES, NO
+from cellprofiler.gui.help import USING_METADATA_HELP_REF
 
 logger = logging.getLogger(__name__)
 C_ANY = "Flag if any fail"
