+++ cellprofiler/rules.py
